fix: avoid pause after vim.notify
The vim.notify calls were moved to the DefaultRunner. As the
orchestrator, it seems more fitting for the DefultRunner to send the
final notifications.

Also, it seems vim.notify causes a pause that prevents us from opening
the quickfix window immediately after a build. So we wrap the call in a
vim.schedule.